+++ src/main/java/com/xawl/Pojo/Testwork.java
@@ -2,48 +2,48 @@ package com.xawl.Pojo;
 
 public class Testwork {
 
-  private long id;
-  private long uid;
-  private long type;
-  private long num;
+  private int id;
+  private int uid;
+  private int type;
+  private int num;
   private String lnames;
-  private long classhours;
-  private long pass;
+  private int classhours;
+  private int pass;
   private java.sql.Timestamp starteddate;
 
 
-  public long getId() {
+  public int getId() {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(int id) {
     this.id = id;
   }
 
 
-  public long getUid() {
+  public int getUid() {
     return uid;
   }
 
-  public void setUid(long uid) {
+  public void setUid(int uid) {
     this.uid = uid;
   }
 
 
-  public long getType() {
+  public int getType() {
     return type;
   }
 
-  public void setType(long type) {
+  public void setType(int type) {
     this.type = type;
   }
 
 
-  public long getNum() {
+  public int getNum() {
     return num;
   }
 
-  public void setNum(long num) {
+  public void setNum(int num) {
     this.num = num;
   }
 
@@ -57,20 +57,20 @@ public class Testwork {
   }
 
 
-  public long getClasshours() {
+  public int getClasshours() {
     return classhours;
   }
 
-  public void setClasshours(long classhours) {
+  public void setClasshours(int classhours) {
     this.classhours = classhours;
   }
 
 
-  public long getPass() {
+  public int getPass() {
     return pass;
   }
 
-  public void setPass(long pass) {
+  public void setPass(int pass) {
     this.pass = pass;
   }
